finish implementing PolyTree, remove old PolyGon
